Add GHE support
Optionally adds GHE support. Bumped the ancient go-github client to get
access to the NewEnterpriseClient func. Some methods like IsTeamMember
have been deprecated[1]. Note that the new methods expect a
context.Context to be passed around. Not particularly useful for what
we are doing here. go-github now uses int64 rather than int for
responses, so updated.

At this point, the thing builds.

[1]
https://github.com/google/go-github/issues/717

+++ cmd/root.go
@@ -87,7 +87,7 @@ Config:
 			GithubAPIToken:     viper.GetString("github_api_token"),
 			GithubOrganization: viper.GetString("github_organization"),
 			GithubTeamName:     viper.GetString("github_team"),
-			GithubTeamID:       viper.GetInt("github_team_id"),
+			GithubTeamID:       viper.GetInt64("github_team_id"),
 
 			EtcdEndpoints: fixStringSlice(viper.GetString("etcd_endpoint")),
 			EtcdPrefix:    viper.GetString("etcd_prefix"),
